+++ Source/DafnyCore/AST/Attributes.cs
@@ -4,7 +4,6 @@ using System.Collections.Generic;
 using System.Diagnostics.Contracts;
 using System.Linq;
 using System.Numerics;
-using JetBrains.Annotations;
 
 namespace Microsoft.Dafny;
 
@@ -105,7 +104,7 @@ public class Attributes : NodeWithComputedRange, ICanFormat {
   /// Returns first occurrence of an attribute named <c>nm</c>, or <c>null</c> if there is no such
   /// attribute.
   /// </summary>
-  [System.Diagnostics.Contracts.Pure]
+  [Pure]
   public static Attributes? Find(Attributes? attrs, string nm) {
     Contract.Requires(nm != null);
     return attrs.AsEnumerable().FirstOrDefault(attr => attr.Name == nm);
@@ -119,7 +118,7 @@ public class Attributes : NodeWithComputedRange, ICanFormat {
   /// This method does NOT use type information of the attribute arguments, so it can safely
   /// be called very early during resolution before types are available and names have been resolved.
   /// </summary>
-  [System.Diagnostics.Contracts.Pure]
+  [Pure]
   public static bool ContainsBool(Attributes? attrs, string nm, ref bool value) {
     Contract.Requires(nm != null);
     var attr = attrs.AsEnumerable().FirstOrDefault(attr => attr.Name == nm);
