+++ Source/DafnyCore/AST/Modules/FileModuleDefinition.cs
@@ -3,14 +3,6 @@ using System.Linq;
 
 namespace Microsoft.Dafny;
 
-public class FilesContainer(List<FileStart> files) {
-  public List<FileStart> Files { get; } = files;
-}
-
-public class FileStart(string uri, List<TopLevelDecl> topLevelDecls) {
-  public string Uri { get; } = uri;
-  public List<TopLevelDecl> TopLevelDecls { get; } = topLevelDecls;
-}
 
 
 /// <summary>
